ARFHTR nodes merged
