+++ src/skmultiflow/trees/nodes/__init__.py
@@ -19,8 +19,7 @@ from .lc_hoeffding_tree_classifier_nodes import LCActiveLearningNode, LCInactive
     LCLearningNodeNB, LCLearningNodeNBA
 from .hoeffding_tree_regressor_nodes import ActiveLearningNodeForRegression, \
     ActiveLearningNodePerceptron, InactiveLearningNodeForRegression, InactiveLearningNodePerceptron
-from .random_learning_node_for_regression import RandomLearningNodeForRegression
-from .random_learning_node_perceptron import RandomLearningNodePerceptron
+from .arf_htr_nodes import RandomLearningNodeForRegression, RandomLearningNodePerceptron
 from .hoeffding_adaptive_tree_regressor_nodes import AdaSplitNodeForRegression, \
     AdaLearningNodeForRegression
 from .active_learning_node_for_regression_multi_target import \
